refactor(server): address review feedback on system settings
- Move AgentTraceRecording/AgentAssistantEnabled GraphQL types from
  queries.py to system_settings_mutations.py (only used there)
- Drop redundant _viewer_user_id helper in favor of the existing
  context.user_id (same semantics; canonical helper used elsewhere)
- Rename local var s -> setting for readability
- Drop commented-out disable block from evals/pxi README
